Cosmetic changes and optimalization

+++ include/memory_game.h
@@ -3,6 +3,6 @@
 
 #include <Adafruit_SH110X.h>
 
-void memoryGame(const int buttonPins[], const int ledPins[], Adafruit_SH1106G& display);
+void memoryGame(const uint8_t buttonPins[], const uint8_t ledPins[], Adafruit_SH1106G& display);
 
 #endif
